[#91] Went too far with the `enum class` thing. Undoing some of the changes
* `enum class` will be relegated to abstract concepts which are never or infrequently
  converted to their ordinal values. "color"s and "mode"s of operation are often good
  examples of this. "Mode" would be a bad example if the ordinal values represent
  specific bits in a register which configures the mode of operation

+++ PropWare/sensor/distance/ping.h
@@ -129,11 +129,9 @@ class Ping {
             this->m_trigger.clear();
 
             this->m_echo.set_dir_in();
-            waitpeq(static_cast<uint32_t>(this->m_echo.get_mask()),
-                    static_cast<uint32_t>(this->m_echo.get_mask()));
+            waitpeq(this->m_echo.get_mask(), this->m_echo.get_mask());
             const uint32_t start = CNT;
-            waitpne(static_cast<uint32_t>(this->m_echo.get_mask()),
-                    static_cast<uint32_t>(this->m_echo.get_mask()));
+            waitpne(this->m_echo.get_mask(), this->m_echo.get_mask());
             return CNT - start;
         }
